requirements.txt
gpytorch~=1.6.0
tqdm~=4.64.1
pandas~=1.4.2
numpy~=1.23.1
pytorch
pydap
setuptools~=63.4.1
matplotlib~=3.5.1
scikit-learn~=1.1.2
scikit-image~=0.19.3